WalletPage-fix-separated tables into component
